Refactor SpringGuildService to use new OperationResult

+++ src/main/kotlin/uk/me/danielharman/kotlinspringbot/listeners/GuildMessageListener.kt
@@ -14,6 +14,7 @@ import net.dv8tion.jda.api.exceptions.ErrorResponseException
 import net.dv8tion.jda.api.hooks.ListenerAdapter
 import org.slf4j.Logger
 import org.slf4j.LoggerFactory
+import org.springframework.http.ResponseEntity
 import org.springframework.stereotype.Component
 import uk.me.danielharman.kotlinspringbot.KotlinBotProperties
 import uk.me.danielharman.kotlinspringbot.helpers.EmojiCodes
@@ -22,6 +23,8 @@ import uk.me.danielharman.kotlinspringbot.models.Meme
 import uk.me.danielharman.kotlinspringbot.factories.ModeratorCommandFactory
 import uk.me.danielharman.kotlinspringbot.factories.CommandFactory
 import uk.me.danielharman.kotlinspringbot.factories.VoiceCommandFactory
+import uk.me.danielharman.kotlinspringbot.helpers.Failure
+import uk.me.danielharman.kotlinspringbot.helpers.Success
 import uk.me.danielharman.kotlinspringbot.services.GuildService
 import uk.me.danielharman.kotlinspringbot.services.MemeService
 import java.util.regex.Pattern
@@ -47,11 +50,12 @@ class GuildMessageListener(
     override fun onGuildJoin(event: GuildJoinEvent) {
         logger.info("Joined guild ${event.guild.name}")
 
-        val guild = guildService.getGuild(event.guild.id)
-        if (guild == null || guild.privilegedUsers.isEmpty()) {
+        val getGuild = guildService.getGuild(event.guild.id)
+
+        if ((getGuild is Failure) || (getGuild as Success).value.privilegedUsers.isEmpty()) {
             val owner = event.guild.retrieveOwner().complete()
             logger.info("Adding ${owner.nickname} as admin of guild")
-            guildService.addPrivileged(event.guild.id, owner.id)
+            guildService.addModerator(event.guild.id, owner.id)
         }
 
         val defaultChannel = event.guild.defaultChannel ?: return
@@ -69,7 +73,13 @@ class GuildMessageListener(
 
         if (event.reactionEmote.isEmoji) {
             val emoji = event.reactionEmote.asCodepoints
-            val guild = guildService.getGuild(event.guild.id) ?: return
+
+            val getGuild = guildService.getGuild(event.guild.id)
+            if (getGuild is Failure) {
+                logger.error("onMessageReactionAdd: ${getGuild.reason}")
+                return
+            }
+            val guild = (getGuild as Success).value
 
             if (guild.memeChannels.contains(event.channel.id)) {
 
@@ -128,7 +138,12 @@ class GuildMessageListener(
 
         if (event.reactionEmote.isEmoji) {
             val emoji = event.reactionEmote.asCodepoints
-            val guild = guildService.getGuild(event.guild.id) ?: return
+            val getGuild = guildService.getGuild(event.guild.id)
+            if (getGuild is Failure) {
+                logger.error("onMessageReactionRemove: ${getGuild.reason}")
+                return
+            }
+            val guild = (getGuild as Success).value
 
             if (guild.memeChannels.contains(event.channel.id)) {
 
@@ -150,7 +165,12 @@ class GuildMessageListener(
 
     override fun onMessageDelete(event: MessageDeleteEvent) {
 
-        val guild = guildService.getGuild(event.guild.id) ?: return
+        val getGuild = guildService.getGuild(event.guild.id)
+        if (getGuild is Failure) {
+            logger.error("onMessageDelete: ${getGuild.reason}")
+            return
+        }
+        val guild = (getGuild as Success).value
 
         if (guild.memeChannels.contains(event.channel.id)) {
             memeService.deleteMeme(guild.guildId, event.messageId)
@@ -163,6 +183,7 @@ class GuildMessageListener(
         val message = event.message
         val guild = event.guild
         val member: Member?
+
         try {
             member = guild.retrieveMember(author).complete()
         } catch (e: ErrorResponseException) {
@@ -175,7 +196,13 @@ class GuildMessageListener(
         if (author.isBot)
             return
 
-        val isDeafened = guildService.getDeafenedChannels(guild.id).contains(event.channel.id)
+        val getDeafenedChannels = guildService.getDeafenedChannels(guild.id)
+        var deafenedChannels = listOf<String>()
+        if(getDeafenedChannels is Success){
+            deafenedChannels = getDeafenedChannels.value
+        }
+
+        val isDeafened = deafenedChannels.contains(event.channel.id)
 
         if (!isDeafened && event.message.isMentioned(event.jda.selfUser, Message.MentionType.USER)) {
             val emotesByName = guild.getEmotesByName("piing", true)
@@ -199,7 +226,13 @@ class GuildMessageListener(
             }
             else -> {
 
-                if (guildService.getMemeChannels(event.guild.id).contains(event.channel.id)) {
+                val getMemeChannels = guildService.getMemeChannels(guild.id)
+                var memeChannels = listOf<String>()
+                if(getMemeChannels is Success){
+                    memeChannels = getMemeChannels.value
+                }
+
+                if (memeChannels.contains(event.channel.id)) {
                     createMeme(event.message, event.guild.id, event.author.id, event.channel.id)
                 }
 
@@ -305,9 +338,9 @@ class GuildMessageListener(
         val cmd = event.message.contentStripped.split(" ")[0].removePrefix(properties.privilegedCommandPrefix)
         val channel = event.channel
 
-        if (event.author.id != properties.primaryPrivilegedUserId
-            && !guildService.isPrivileged(event.guild.id, event.author.id)
-        ) {
+        val isModerator = guildService.isModerator(event.guild.id, event.author.id)
+
+        if (event.author.id != properties.primaryPrivilegedUserId && isModerator is Failure) {
             channel.sendMessage("You are not an admin!").queue()
             return
         }
